+++ src/main/aug_jac/direct_aug_jac.h
@@ -1,0 +1,19 @@
+#ifndef SLEQP_DIRECT_AUG_JAC_H
+#define SLEQP_DIRECT_AUG_JAC_H
+
+#include "aug_jac.h"
+#include "fact/fact_qr.h"
+
+/**
+ * Works on the augmented Jacobian
+ * by QR factorizing \f$ A_W \f$
+ *
+ **/
+SLEQP_NODISCARD
+SLEQP_RETCODE
+sleqp_direct_aug_jac_create(SleqpAugJac** star,
+                            SleqpProblem* problem,
+                            SleqpParams* params,
+                            SleqpFactQR* fact);
+
+#endif /* SLEQP_DIRECT_AUG_JAC_H */