+++ src/main/fact/fact_ma86.c
@@ -281,16 +281,6 @@ ma86_data_solution(void* fact_data,
 }
 
 static SLEQP_RETCODE
-ma86_data_condition_estimate(void* fact_data, double* condition_estimate)
-{
-  // MA86Data* ma86_data = (MA86Data*) fact_data;
-
-  (*condition_estimate) = SLEQP_NONE;
-
-  return SLEQP_OKAY;
-}
-
-static SLEQP_RETCODE
 ma86_data_free(void** star)
 {
   MA86Data* ma86_data = (MA86Data*)(*star);
@@ -313,12 +303,11 @@ ma86_data_free(void** star)
 SLEQP_RETCODE
 sleqp_fact_ma86_create(SleqpFact** star, SleqpParams* params)
 {
-  SleqpFactorizationCallbacks callbacks
-    = {.set_matrix         = ma86_data_set_matrix,
-       .solve              = ma86_data_solve,
-       .solution           = ma86_data_solution,
-       .condition_estimate = ma86_data_condition_estimate,
-       .free               = ma86_data_free};
+  SleqpFactCallbacks callbacks = {.set_matrix = ma86_data_set_matrix,
+                                  .solve      = ma86_data_solve,
+                                  .solution   = ma86_data_solution,
+                                  .condition  = NULL,
+                                  .free       = ma86_data_free};
 
   MA86Data* ma86_data;