+++ src/main/lp/lpi.h
@@ -109,9 +109,7 @@ sleqp_lpi_solve_timer(SleqpLPi* lp_interface);
 
 SLEQP_NODISCARD
 SLEQP_RETCODE
-sleqp_lpi_basis_condition_estimate(SleqpLPi* lp_interface,
-                                   bool* exact,
-                                   double* condition);
+sleqp_lpi_basis_cond(SleqpLPi* lp_interface, bool* exact, double* condition);
 
 SLEQP_NODISCARD
 SLEQP_RETCODE