+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added direct QR factorization to solve augmented Jacobian systems
+
 ## [0.1.0] - 2022-07-07
 
 Initial release